Add listPosition property
The listPosition property defines where list should be displayed around
the select main element (at top or at bottom).

'bottom': the list is always displayed at bottom of the element.
'top': the list is always displayed at top of the element.
'auto': the list is displayed at bottom if there is enough space, at
'top' otherwise (this is the default value).