fix double call of getItem
When the dynamic mode is enabled, there were some issues because
getItems is called several times.

It happened when previous request is not fetched yet and current
selected items need to be updated.

Now, a cache system has been added to avoid resending a request if
it has already been called.