fix change static texts.
The change of static strings should not change the texts of existing
instance because it depending on strings, the change won't be really be
displayed on the component (there is no reactivity for this change).

So if the change need to be changed on existing instance , it would be
better to use changeText directly on them.