Add "open" attribute
The "open" attribute allows to create a component which is open at
start.
But it also allows to open a component programatically (depending on a
state).
To do so, it would need to keep track of the internal open state.

A method toggleOpen() has also been added to switch the "open" state
easily. The wanted state can be forced by setting the only argument to
the desired state value.
If the state is already to the given value, no changes are triggered.